"Added password parameter in creating new users
Added password parameter to the 'new' function in 'src/controllers/users/new/main.js' for creating a new user. This allows the creation of new users with a password, increasing account security. Also modified the POST method of 'userRouter' in 'src/routes/user/new/main.js' to handle correctly the received username and password, check for their validity, hash the password, and create a new user if validation is successful. This change validates the input data from the client ensuring both robustness and security in new user creation."