Added route initialization for tokens, users and messages
This change includes the addition of route initialization in individual router JavaScript files for tokens, users and messages respectively. This was done to improve the organization and readability of the code and make it easier to manage and maintain. It will now be easier to add or modify features related to specific routes.