Add login parameter to getToken function
Adjusted the getToken function within the JWT module to now accept a 4th parameter representing login data. This additional parameter allows for the function to generate a new token for an existing user logging in, rather than only for new users as before. The goal is to allow more flexibility in token generation and enhance security by replacing tokens upon each successful login.