Improve findAndDeleteTokens function efficiency and readability
Refactoring on the findAndDeleteTokens function has been carried out. Changed function name to camel case and restructured the logic to re-use the MongoDB client. This improves the function's efficiency by reducing the number of created clients and connections to the database. Also, updated the token deletion for both access and refresh token collections, now done separately for each, making the code more readable. In case of error, it now logs the error before throwing it, which aids debugging. Also, the client closure is ensured inside a "finally" block.