Implement SameSite setting in cookie options and refine accessToken type
In the cookie_settings file, a SameSite setting was added to the CookieOptions  for improving security by preventing CSRF attacks. In the check_token file, the type of the accessToken parameter in the `checkAccessToken` function has been refined from `any` to `string` to ensure type safety and improve the reliability of the code.